Move version logic to kolibri utils.
